+++ backend/api/article/controllers/article.js
@@ -7,7 +7,7 @@ module.exports = {
             const { data, files } = parseMultipartData(ctx);
             entity = await strapi.services.comment.create(data, { files });
         } else {
-            ctx.request.body.author = 2;
+            ctx.request.body.user = ctx.state.user.id;
             ctx.request.body.article = ctx.params.id;
             entity = await strapi.services.comment.create(ctx.request.body);
         }
