+++ backend/api/author/models/author.settings.json
@@ -34,10 +34,6 @@
     "articles": {
       "via": "author",
       "collection": "article"
-    },
-    "comments": {
-      "via": "author",
-      "collection": "comment"
     }
   }
 }
