+++ backend/api/comment/models/comment.settings.json
@@ -16,13 +16,13 @@
       "type": "text",
       "required": true
     },
-    "author": {
-      "model": "author",
-      "via": "comments"
-    },
     "article": {
       "via": "comments",
       "model": "article"
+    },
+    "user": {
+      "plugin": "users-permissions",
+      "model": "user"
     }
   }
 }
